# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于Go+Vue实现的openLDAP后台管理项目",
        "title": "Go Ldap Admin",
        "termsOfService": "https://github.com/eryajf/go-ldap-admin",
        "contact": {
            "name": "项目作者：二丫讲梵 、 swagger作者：南宫乘风",
            "url": "https://github.com/eryajf/go-ldap-admin",
            "email": "https://github.com/eryajf/go-ldap-admin"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8888",
    "basePath": "/api",
    "paths": {
...
        "/log/operation/clean": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "操作日志管理"
                ],
                "summary": "清空操作日志记录",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/log/operation/delete": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "操作日志管理"
                ],
                "summary": "删除操作日志记录",
                "parameters": [
                    {
                        "description": "删除日志的ID",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.OperationLogDeleteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/log/operation/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "操作日志管理"
                ],
                "summary": "获取操作日志记录列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "IP地址",
                        "name": "ip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "路径",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "方法",
                        "name": "method",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态码",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/menu/access/tree": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "获取用户菜单树",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分组ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/menu/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "新建菜单",
                "parameters": [
                    {
                        "description": "新建菜单",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MenuAddReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/menu/delete": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "删除菜单",
                "parameters": [
                    {
                        "description": "删除菜单",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MenuDeleteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/menu/tree": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "获取菜单树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/menu/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "更新菜单",
                "parameters": [
                    {
                        "description": "更新菜单",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MenuUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.MenuAddReq": {
            "type": "object",
            "required": [
                "component",
                "name",
                "path",
                "title"
            ],
            "properties": {
                "activeMenu": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 0
                },
                "alwaysShow": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "breadcrumb": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "component": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "hidden": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "icon": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 0
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "noCache": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "parentId": {
                    "type": "integer"
                },
                "path": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "redirect": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 0
                },
                "sort": {
                    "type": "integer",
                    "maximum": 999,
                    "minimum": 1
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "title": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                }
            }
        },
        "request.MenuDeleteReq": {
            "type": "object",
            "required": [
                "menuIds"
            ],
            "properties": {
                "menuIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "request.MenuUpdateReq": {
            "type": "object",
            "required": [
                "id",
                "name",
                "path",
                "title"
            ],
            "properties": {
                "activeMenu": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 0
                },
                "alwaysShow": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "breadcrumb": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "component": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 0
                },
                "hidden": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "icon": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 0
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "noCache": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "parentId": {
                    "type": "integer",
                    "minimum": 0
                },
                "path": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "redirect": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 0
                },
                "sort": {
                    "type": "integer",
                    "maximum": 999,
                    "minimum": 1
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "title": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                }
            }
        },
        "request.OperationLogDeleteReq": {
            "type": "object",
            "required": [
                "operationLogIds"
            ],
            "properties": {
                "operationLogIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "response.ResponseBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}